Add region list API handler
I'm using a github API crate I found as inspiration, but essentially
I will pass in my main struct into the different handlers so that
they will be able to call the specific methods on it. What I might
end up doing is passing in some other struct just to get another
layer of indirection, but what I have is fine for now.

I added lifetimes to everything as well for ease of use.