Add adjacent regions
This adds in the ability to look up adjacent regions for a given region
code. However, this API is fairly limited.

Doing this also exposed that I wasn't handling an error response from
ebird appropriately, and was just dying on the parse. Instead of dying
on the parse, I'm checking first to see. I decided to not use the
error_from_status that is available in reqwest because I want to just
provide the response body and the text.

I'm not parsing the text of the error because I can't guarantee it
being valid JSON.